Modify layout parameter
