+++ cocos/ui/UILayout.h
@@ -26,6 +26,7 @@ THE SOFTWARE.
 #define __LAYOUT_H__
 
 #include "ui/UIWidget.h"
+#include "ui/CCLayoutProtocol.h"
 
 NS_CC_BEGIN
 
@@ -38,14 +39,6 @@ typedef enum
     LAYOUT_COLOR_GRADIENT
 }LayoutBackGroundColorType;
 
-typedef enum
-{
-    LAYOUT_ABSOLUTE,
-    LAYOUT_LINEAR_VERTICAL,
-    LAYOUT_LINEAR_HORIZONTAL,
-    LAYOUT_RELATIVE
-}LayoutType;
-
 typedef enum {
     LAYOUT_CLIPPING_STENCIL,
     LAYOUT_CLIPPING_SCISSOR
@@ -57,7 +50,7 @@ typedef enum {
  */
 class LayoutExecutant;
     
-class Layout : public Widget
+class Layout : public Widget, public LayoutProtocol
 {
     
     DECLARE_CLASS_GUI_INFO
@@ -315,7 +308,7 @@ protected:
     //init background image renderer.
     void addBackGroundImage();
     
-    void supplyTheLayoutParameterLackToChild(Widget* child);
+    void supplyTheLayoutParameterLackToChild(LayoutParameterProtocol* child);
     virtual Widget* createCloneInstance() override;
     virtual void copySpecialProperties(Widget* model) override;
     virtual void copyClonedWidgetChildren(Widget* model) override;
@@ -460,7 +453,6 @@ protected:
     Vector2 _alongVector;
     GLubyte _cOpacity;
     Size _backGroundImageTextureSize;
-    LayoutType _layoutType;
     LayoutClippingType _clippingType;
     DrawNode* _clippingStencil;
     bool _scissorRectDirty;
@@ -489,7 +481,6 @@ protected:
     Color3B _backGroundImageColor;
     GLubyte _backGroundImageOpacity;
     
-    LayoutExecutant* _curLayoutExecutant;
     
     GLint _mask_layer_le;
     GroupCommand _groupCommand;
